requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d==0.14.1
opencv-python
pyyaml
scipy
tensorboard
timm==0.4.5
tqdm
transforms3d
protobuf==3.20.1
torch_kmeans
geomdl
bezier
scikit-learn